Added a Show View
- Used React Router to easily generate routes and help with navigation
  between the index and show views
- Refactored SocialMediaPostsTable into a compoenent to clean things up a
  bit
- Added a new SocialMediaPostDetail component to display each social
  media posts show/detail view
